Main menu work
Fixed a few formatting issues with level select menu, added a return button, and made the buttons functional.
The buttons on the level select are appropriately set to greyed out/uninteractable when they need to be.
High score text is now displayed under completed levels
Start button now changes to a continue button if appropriate and loads the next unbeaten level
Moved level select onto the same scene as the main menu