Even more bugfixes
Fixed an error stemming from our Cell tag somehow having gotten deleted. Likely was what was causing the rapid duplication, as the scripts would stop working.
Additionally, fixed an issue where loading a new level after beating or losing a previous one would cause the same win/loss outcome immediately as soon as the new level checks for win loss. This was due to not unsubscribing the script in ondestroy.